requirements2.txt
accelerate==1.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
bitsandbytes==0.47.0
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.2
cycler==0.12.1
datasets==3.1.0
dill==0.3.8
einops==0.8.1
filelock==3.19.1
flash-attn==2.7.4.post1
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2024.9.0
hf-xet==1.1.8
huggingface-hub==0.34.4
idna==3.10
jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.9
lightning-utilities==0.15.2
markupsafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
networkx==3.4.2
numpy==1.26.4
pandas==2.3.1
pillow==11.3.0
propcache==0.3.2
pyarrow==21.0.0
pyparsing==3.2.3
pytz==2025.2
pyyaml==6.0.2
regex==2025.7.34
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.4
torchmetrics==1.8.1
tqdm==4.67.1
transformers==4.49.0
triton==3.1.0
tzdata==2025.2
urllib3==2.5.0
xrfm==0.3.0
xxhash==3.5.0
yarl==1.20.1
